test_helper.rb: factor out upload test helpers.
